Remove some warnings from GCC
 - change old casting to new casting (functionnal casting)
 - initialize every member on object creation
 - try to keep consitancy in type (using float in whole Vvector)
 - use nullptr instead of NULL
 - move some member to avoid losing space
 - virtualize destructor of Object
 - use a delta to compare Vvector instead of pure equality

+++ Simulator.h
@@ -29,7 +29,7 @@ class Simulator {
 		/**
 		 * Number of ticks
 		 */
-		unsigned long long tickCounter;
+		unsigned long tickCounter;
 
 		/**
 		 * Last \ref actualPhysic we reported
